NEED TO FIX THE JUNK ON THE BYTES ARRAY IN THE ENCODER (ZEROBYTES)

+++ src/main/java/bgu/spl/net/impl/BGRSServer/users/Student.java
@@ -15,7 +15,7 @@ public class Student extends User{
     }
 
     public void unRegisterToCourse(short courseNum){
-        myRegisteredCourses.remove(courseNum);
+        myRegisteredCourses.remove(Short.valueOf(courseNum));
     }
 
     public ArrayList<Short> getMyRegisteredCourses() {
